refactor[#17]: test 패키지 구조 변경
