--- backend/src/main/resources/templates/reviews/createForm.html
@@ -1,34 +1,0 @@
-
-<!DOCTYPE HTML>
-<html xmlns:th="http://www.thymeleaf.org">
-<!--<head th:replace="fragments/header :: header" /> -->
-<style>
-    .fieldError {
-        border-color: #bd2130;
-    }
-</style>
-<body>
-<div class="container">
-    <!--<div th:replace="fragments/bodyHeader :: bodyHeader" />-->
-    <form role="form" action="/admin/reviews/new" th:object="${adminReviewPostRequest}" method="post">
-        <div class="form-group">
-            <label th:for="title">title</label>
-            <input type="text" th:field="*{title}" class="form-control" placeholder="제목을 입력하세요"
-                   th:class="${#fields.hasErrors('title')}? 'form-control fieldError' : 'form-control'">
-            <p th:if="${#fields.hasErrors('title')}"
-               th:errors="*{title}">Incorrect date</p>
-        </div>
-        <div class="form-group">
-            <label th:for="content">content</label>
-            <input type="text" th:field="*{content}" class="form-control"  placeholder="내용을 입력하세요"> </div>
-        <div class="form-group">
-            <label th:for="tip">tip</label>
-            <input type="text" th:field="*{tip}" class="form-control" placeholder="팁 입력하세요">
-        </div>
-        <button type="submit" class="btn btn-primary">Submit</button>
-    </form>
-    <br/>
-    <!--<div th:replace="fragments/footer :: footer" />-->
-</div> <!-- /container -->
-</body>
-</html>
