+++ backend/src/test/java/com/handong/rebon/unit/review/domain/content/ReviewContentTest.java
@@ -1,7 +1,8 @@
-package com.handong.rebon.review.domain.content;
+package com.handong.rebon.unit.review.domain.content;
 
 import com.handong.rebon.exception.review.ReviewContentFormatException;
 import com.handong.rebon.exception.review.ReviewTitleFormatException;
+import com.handong.rebon.review.domain.content.ReviewContent;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
